Consolidate TrackingPoint code.
